Current status: In conversion
Currently implemented `.getspots` function, which merges functions of
`getspots` and `sortspots` together. Hotspots are now sourced using the
`krpano.get` function, instead of directly scraping XML file.

The same applies for hotspot actions, which now simply appends `jscall`
to the end of the hotspots actions. This fixes the mysterious bug of
pano spots not zooming if they had onclick events.

Still to be done:
 - Implement a history manager again
 - Implement on-screen drawing
 - Implement history-reviver
 - Implement a better localStorage manager
    - Along with better storage management, also build a '.save' function
      to dump only necessary saved-data.
    - Implement a per-tour save setup and cleaner, so it doesn't wipe
      /all/ localStorage from the domain. **THIS IS URGENT
    - Possibly add in a 'secret' way to flush the storage? e.g. hotkeys
      or click and hold on something for n seconds?
  - Investigate minifying JS files and using those for permanent builds.